flask-server/requirements.txt
tensorflow==2.17.0
pandas==2.2.2
flask==3.0.3
numpy==1.26.4
pillow==10.4.0
requests==2.32.3
streamlit==1.37.0
protobuf==4.25.4
translate==3.6.1
wheel==0.43.0
Keras-Preprocessing==1.1.2